% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuctions.R
\name{iNEXT.dare}
\alias{iNEXT.dare}
\title{A main Function}
\usage{
iNEXT.dare(
  x,
  rho,
  q = 0,
  datatype = "abundance",
  endpoint = NULL,
  knots = 40,
  se = TRUE,
  conf = 0.95,
  nboot = 50
)
}
\arguments{
\item{x}{A list or a data.frame.}

\item{q}{a paremeter of hill numbers}
}
\description{
A main Function
}
