% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfunction.R, R/iNEXTdare.R
\name{iNEXTdare}
\alias{iNEXTdare}
\title{A main Function}
\usage{
iNEXTdare(
  x,
  rho,
  q = 0,
  datatype = "abundance",
  endpoint = NULL,
  knots = 40,
  se = TRUE,
  conf = 0.95,
  nboot = 50
)

iNEXTdare(
  x,
  rho,
  q = 0,
  datatype = "abundance",
  endpoint = NULL,
  knots = 40,
  se = TRUE,
  conf = 0.95,
  nboot = 50
)
}
\arguments{
\item{x}{A list or a data.frame.}

\item{q}{a paremeter of hill numbers}

\item{datatype}{datatype can abundance and incidence_freq}
}
\description{
A main Function

A main Function
}
