% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfunction.R
\name{iNEXTdare}
\alias{iNEXTdare}
\title{Main Functions
\code{iNEXT_dare}: compute the rarefaction and extrapolation of diveristy without.}
\usage{
iNEXTdare(
  x,
  rho,
  q = 0,
  datatype = "abundance",
  endpoint = NULL,
  knots = 40,
  se = TRUE,
  conf = 0.95,
  nboot = 50
)
}
\arguments{
\item{x}{a list consist of N data.frame/matrix describing species-by-assemblage/plot abundance. Note that
the species in each element must exactly match including specpes order. Use \code{data(abundata)} to see data example.}

\item{rho}{a numeric value of the fraction of sample size, which should be between 0 and 1.}

\item{q}{a numeric value or vector specifying the diversity order of Hill number.}

\item{datatype}{data type of input data: individual-based abundance data (datatype = "abundance"),
sampling-unit-based incidence frequencies data (datatype = "incidence_freq")}

\item{endpoint}{an integer specifying the sample size that is the endpoint for rarefaction/extrapolation.
If NULL, then endpoint = double reference sample size.}

\item{knots}{an integer specifying the number of equally-spaced knots (say K, default is 40) between size 1 and the endpoint;
each knot represents a particular sample size for which diversity estimate will be calculated. If the endpoint is smaller than the reference sample size,}

\item{se}{a logical variable to calculate the bootstrap standard error and conf confidence interval.}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.}

\item{nboot}{an integer specifying the number of replications.}

\item{size}{an integer vector of sample sizes (number of individuals or sampling units) for which diversity estimates will be computed.
If NULL, then diversity estimates will be computed for those sample sizes determined by the specified/default endpoint and knots.}
}
\description{
Main Functions
\code{iNEXT_dare}: compute the rarefaction and extrapolation of diveristy without.
}
\examples{
data(abudata)
out <- iNEXTdare(x = abudata, q = c(0,1,2), datatype = "abundance",se = TRUE)
}
